requirements.txt
openai>=1.0.0
SpeechRecognition==3.14.5
gTTS==2.5.4
pygame>=2.5.0
pyttsx3>=2.90
PyAudio==0.2.14
python-dotenv
requests>=2.31.0
numpy
perplexityai